Added expiry date check

+++ UrlShortenerApi/Models/ShortenRequest.cs
@@ -3,5 +3,5 @@ namespace UrlShortenerApi.Models;
 public class ShortenRequest
 {
     public string OriginalUrl { get; set; }
-    public int? ExpiresInDays { get; set; }
+    public double? ExpiresInDays { get; set; }
 }
